requirements.txt
streamlit
anthropic
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
tiktoken
openai